Prototype for the Subscription
right after solving the problem with cookies ( when I would enter in number of fields info, I would do the subscribe. After that in the navbar I would do Update, enter info in one of fields - after doing subscribe that one field would be filled in and the rest of them too - this is the SOLUTION for this ). Now I have to implement to the other forms.